move the secret settings/config to a separate class so that we can still monkeypatch the other env var settings
