+++ api/app/config.py
@@ -8,10 +8,6 @@ TMDB_API_URL = "https://api.themoviedb.org/3"
 
 class Settings(BaseSettings):
 
-    secret: str = Field(
-        ..., env="SECRET_KEY", description="A unique unpredictable value"
-    )
-
     tmdb_api_url: str = TMDB_API_URL
     tmdb_api_key: str = Field(
         ...,
@@ -21,7 +17,7 @@ class Settings(BaseSettings):
     tmdb_base_path: HttpUrl | None = None
 
     @validator("tmdb_base_path", always=True)
-    def set_tmdb_base_path(cls, v, values):
+    def set_tmdb_base_path(cls, value: str, values: dict):
         """Assign the tmdb_base_path"""
 
         # we use a validator here to "compute" tmdb_base_path, which requires tmdb_api_key
@@ -30,11 +26,27 @@ class Settings(BaseSettings):
         # get configuration for poster base_url (e.g.: https://image.tmdb.org/t/p/)
         #  https://api.themoviedb.org/3/configuration?api_key=<<api_key>>
 
-        tmdb_config_url = f"{values['tmdb_api_url']}/configuration"
-        config_resp = httpx.get(
-            tmdb_config_url, params={"api_key": values["tmdb_api_key"]}
-        )
-        return config_resp.json()["images"]["secure_base_url"]
+        if value is None:
+            tmdb_config_url = f"{values['tmdb_api_url']}/configuration"
+            config_resp = httpx.get(
+                tmdb_config_url, params={"api_key": values["tmdb_api_key"]}
+            )
+            value = config_resp.json()["images"]["secure_base_url"]
+        return value
+
+    class Config:
+        env_file = ".env"
+        env_file_encoding = "utf-8"
+
+
+class Secret(BaseSettings):
+    """This settings config needs to be loaded at import time
+    since it's used to set up the login manager decorator so for testing,
+    we've split it out into a separate class"""
+
+    secret: str = Field(
+        ..., env="SECRET_KEY", description="A unique unpredictable value"
+    )
 
     class Config:
         env_file = ".env"
@@ -48,3 +60,8 @@ def get_settings():
     Note: we use @lru_cache to avoid calling the configuration endpoint over and over
     """
     return Settings()
+
+
+@lru_cache
+def get_secret():
+    return Secret().secret
