+++ api/requirements.txt
@@ -21,6 +21,10 @@ certifi==2022.9.24
     #   httpx
 click==8.1.3
     # via uvicorn
+dnspython==2.2.1
+    # via email-validator
+email-validator==1.3.0
+    # via movies_from_a_hat (pyproject.toml)
 fastapi==0.88.0
     # via
     #   fastapi-login
@@ -42,6 +46,7 @@ httpx==0.23.1
 idna==3.4
     # via
     #   anyio
+    #   email-validator
     #   rfc3986
 loguru==0.6.0
     # via movies_from_a_hat (pyproject.toml)
