+++ api/requirements_dev.txt
@@ -26,6 +26,10 @@ charset-normalizer==2.1.1
     # via requests
 click==8.1.3
     # via uvicorn
+dnspython==2.2.1
+    # via email-validator
+email-validator==1.3.0
+    # via movies_from_a_hat (pyproject.toml)
 faker==15.3.4
     # via movies_from_a_hat (pyproject.toml)
 fastapi==0.88.0
@@ -51,6 +55,7 @@ httpx==0.23.1
 idna==3.4
     # via
     #   anyio
+    #   email-validator
     #   requests
     #   rfc3986
 iniconfig==1.1.1
