+++ api/tests/conftest.py
@@ -1,25 +1,22 @@
 import json
-import pathlib
+import os
+from pathlib import Path
 from unittest.mock import patch
 
 import pytest
 import respx
 from _pytest.logging import LogCaptureFixture
 from faker import Faker
-from fastapi.testclient import TestClient
 from httpx import Response
 from loguru import logger
 from respx.patterns import M
 from sqlalchemy.ext.asyncio import create_async_engine
 from sqlalchemy.future import Engine
 from sqlalchemy.orm import sessionmaker
-from sqlmodel import Session
 from sqlmodel.ext.asyncio.session import AsyncSession
 from sqlmodel.pool import StaticPool
 
 from app import config
-from app.api import app
-from app.db import get_session
 
 
 @pytest.fixture
@@ -36,8 +33,15 @@ def caplog(caplog: LogCaptureFixture):
     logger.remove(handler_id)
 
 
+def pytest_configure(config):
+    # in api, secret config needs to be present at import time because it's used
+    # to create the login manager decorator
+    # pytest_configure is a hook that runs at pytest startup, before anything else
+    os.environ["SECRET_KEY"] = "secret"
+
+
 @pytest.fixture(autouse=True)
-def monkeypatch_settings_env_vars(monkeypatch):
+def patch_env_var(monkeypatch: pytest.MonkeyPatch):
     monkeypatch.setenv("TMDB_API_TOKEN", "TESTING")
 
 
@@ -75,36 +79,32 @@ async def session_fixture(engine: Engine):
         yield session
 
 
-@pytest.fixture(name="client")
-async def client_fixture(session: Session):
-    """Create the test client
+@pytest.fixture
+async def mocked_TMDB_config_req(respx_mock):
 
-    Overrides the session dependency in our endpoints to use this session instead
+    mocked_tmdb_config_data = {
+        "images": {
+            "base_url": "http://image.tmdb.org/t/p/",
+            "secure_base_url": "https://image.tmdb.org/t/p/",
+            "backdrop_sizes": ["w300", "w780", "w1280", "original"],
+            "logo_sizes": ["w45", "w92", "w154", "w185", "w300", "w500", "original"],
+            "poster_sizes": ["w92", "w154", "w185", "w342", "w500", "w780", "original"],
+            "profile_sizes": ["w45", "w185", "h632", "original"],
+            "still_sizes": ["w92", "w185", "w300", "original"],
+        },
+        # note: there's a list of "change_keys" in the response but we're not using that data
+    }
+    with respx.mock(assert_all_called=False) as respx_mock:
+        respx_mock.get(f"{config.TMDB_API_URL}/configuration").mock(
+            return_value=Response(200, json=mocked_tmdb_config_data)
+        )
 
-    Note that because we create the database tables with a startup event
-    (using create_db_and_tables), this fixture must occur in the parameter list before
-    other database access fixtures.
+        yield respx_mock
 
-    E.g. works:
 
-    def my_test_func(client: TestClient, dude_movie: Movie): ...
-
-    Doesn't work, because dude_movie fixture tries to access database tables:
-
-    def my_test_func(dude_movie: Movie, client: TestClient): ...
-    """
-
-    def get_session_override():
-        return session
-
-    # Set the dependency override in the app.dependency_overrides dictionary.
-    app.dependency_overrides[get_session] = get_session_override
-
-    # context manager runs the app startup/shutdown/lifespan events
-    with TestClient(app) as client:
-        yield client
-
-    app.dependency_overrides.clear()
+@pytest.fixture(name="settings")
+async def settings_fixture(mocked_TMDB_config_req):
+    yield config.get_settings()
 
 
 @pytest.fixture
@@ -139,37 +139,9 @@ async def mocked_TMDB():
 
 
 @pytest.fixture
-async def mocked_TMDB_config_req(respx_mock):
-
-    mocked_tmdb_config_data = {
-        "images": {
-            "base_url": "http://image.tmdb.org/t/p/",
-            "secure_base_url": "https://image.tmdb.org/t/p/",
-            "backdrop_sizes": ["w300", "w780", "w1280", "original"],
-            "logo_sizes": ["w45", "w92", "w154", "w185", "w300", "w500", "original"],
-            "poster_sizes": ["w92", "w154", "w185", "w342", "w500", "w780", "original"],
-            "profile_sizes": ["w45", "w185", "h632", "original"],
-            "still_sizes": ["w92", "w185", "w300", "original"],
-        },
-        # note: there's a list of "change_keys" in the response but we're not using that data
-    }
-    with respx.mock(assert_all_called=False) as respx_mock:
-        respx_mock.get(f"{config.TMDB_API_URL}/configuration").mock(
-            return_value=Response(200, json=mocked_tmdb_config_data)
-        )
-
-        yield respx_mock
-
-
-@pytest.fixture(name="settings")
-async def settings_fixture(mocked_TMDB_config_req):
-    yield config.get_settings()
-
-
-@pytest.fixture
 async def mocked_TMDB_movie_results(request):
     # test data location relative to the test file
-    file = pathlib.Path(request.node.fspath.strpath)
+    file = Path(request.node.fspath.strpath)
     movies_data = {
         "115": json.load(open(file.parent / "test_data" / "115.json")),
         "550": json.load(open(file.parent / "test_data" / "550.json")),
